backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
strawberry-graphql[fastapi]